Consolidate datasets selected by expression hooks
